added all code for genesis validators
