+++ data/sbc/sbcSecrets1.csv
@@ -1,0 +1,100 @@
+https://heyanon.xyz/magic/gwh2qu,Apprentice_0.png
+https://heyanon.xyz/magic/lsctxp,Apprentice_1.png
+https://heyanon.xyz/magic/2u3p56,Apprentice_2.png
+https://heyanon.xyz/magic/phy4az,Apprentice_3.png
+https://heyanon.xyz/magic/5ig62j,Apprentice_4.png
+https://heyanon.xyz/magic/gu4j6z,Apprentice_5.png
+https://heyanon.xyz/magic/vfcyv5,Apprentice_6.png
+https://heyanon.xyz/magic/3oxvh6,Apprentice_7.png
+https://heyanon.xyz/magic/bzutgl,Apprentice_8.png
+https://heyanon.xyz/magic/u8bf4h,Apprentice_9.png
+https://heyanon.xyz/magic/sen5oz,Apprentice_10.png
+https://heyanon.xyz/magic/98avzs,Apprentice_11.png
+https://heyanon.xyz/magic/dqy2r2,Apprentice_12.png
+https://heyanon.xyz/magic/1td4px,Apprentice_13.png
+https://heyanon.xyz/magic/bo284x,Apprentice_14.png
+https://heyanon.xyz/magic/7k1qiq,Apprentice_15.png
+https://heyanon.xyz/magic/wg6nf0,Apprentice_16.png
+https://heyanon.xyz/magic/hobkdz,Apprentice_17.png
+https://heyanon.xyz/magic/q71w3l,Apprentice_18.png
+https://heyanon.xyz/magic/wf20ji,Apprentice_19.png
+https://heyanon.xyz/magic/9xkwsi,Apprentice_20.png
+https://heyanon.xyz/magic/bp0c34,Apprentice_21.png
+https://heyanon.xyz/magic/efmgdy,Apprentice_22.png
+https://heyanon.xyz/magic/vgag27,Apprentice_23.png
+https://heyanon.xyz/magic/x277fg,Apprentice_24.png
+https://heyanon.xyz/magic/xu3c4d,Apprentice_25.png
+https://heyanon.xyz/magic/kdfnj9,Apprentice_26.png
+https://heyanon.xyz/magic/3gz42c,Apprentice_27.png
+https://heyanon.xyz/magic/f9tym5,Apprentice_28.png
+https://heyanon.xyz/magic/16eu1s,Apprentice_29.png
+https://heyanon.xyz/magic/h08gra,Apprentice_30.png
+https://heyanon.xyz/magic/if9gm6,Apprentice_31.png
+https://heyanon.xyz/magic/wzvhpl,Apprentice_32.png
+https://heyanon.xyz/magic/zoqkgp,Apprentice_33.png
+https://heyanon.xyz/magic/31gjj4,Apprentice_34.png
+https://heyanon.xyz/magic/6i6xw3,Apprentice_35.png
+https://heyanon.xyz/magic/jb331a,Apprentice_36.png
+https://heyanon.xyz/magic/45v572,Apprentice_37.png
+https://heyanon.xyz/magic/laaxdk,Apprentice_38.png
+https://heyanon.xyz/magic/sl3x87,Apprentice_39.png
+https://heyanon.xyz/magic/14tcmf,Apprentice_40.png
+https://heyanon.xyz/magic/f6ye27,Apprentice_41.png
+https://heyanon.xyz/magic/0ghkvw,Apprentice_42.png
+https://heyanon.xyz/magic/ulbzmw,Apprentice_43.png
+https://heyanon.xyz/magic/m3aog4,Apprentice_44.png
+https://heyanon.xyz/magic/dksdsm,Apprentice_45.png
+https://heyanon.xyz/magic/tbdj41,Apprentice_46.png
+https://heyanon.xyz/magic/2bgsfl,Apprentice_47.png
+https://heyanon.xyz/magic/i3hs39,Apprentice_48.png
+https://heyanon.xyz/magic/dx9wtw,Apprentice_49.png
+https://heyanon.xyz/magic/bztt7u,Apprentice_50.png
+https://heyanon.xyz/magic/6dwqvn,Apprentice_51.png
+https://heyanon.xyz/magic/o208ez,Apprentice_52.png
+https://heyanon.xyz/magic/fj9y01,Apprentice_53.png
+https://heyanon.xyz/magic/pkj0zk,Apprentice_54.png
+https://heyanon.xyz/magic/h4aw5g,Apprentice_55.png
+https://heyanon.xyz/magic/x49i2n,Apprentice_56.png
+https://heyanon.xyz/magic/r9tkoh,Apprentice_57.png
+https://heyanon.xyz/magic/ymsqjm,Apprentice_58.png
+https://heyanon.xyz/magic/n4ud5g,Apprentice_59.png
+https://heyanon.xyz/magic/sazefy,Apprentice_60.png
+https://heyanon.xyz/magic/0e6s48,Apprentice_61.png
+https://heyanon.xyz/magic/pjjhjh,Apprentice_62.png
+https://heyanon.xyz/magic/jvw4vv,Apprentice_63.png
+https://heyanon.xyz/magic/kfd3sa,Apprentice_64.png
+https://heyanon.xyz/magic/ukdfm4,Apprentice_65.png
+https://heyanon.xyz/magic/xy5ep6,Apprentice_66.png
+https://heyanon.xyz/magic/er2u5b,Apprentice_67.png
+https://heyanon.xyz/magic/azf5zj,Apprentice_68.png
+https://heyanon.xyz/magic/2yqlcy,Apprentice_69.png
+https://heyanon.xyz/magic/5wqzk6,Apprentice_70.png
+https://heyanon.xyz/magic/md7qhj,Apprentice_71.png
+https://heyanon.xyz/magic/f3wgps,Apprentice_72.png
+https://heyanon.xyz/magic/6jgxyy,Apprentice_73.png
+https://heyanon.xyz/magic/gydbfc,Apprentice_74.png
+https://heyanon.xyz/magic/6vy6i4,Apprentice_75.png
+https://heyanon.xyz/magic/an1xk0,Apprentice_76.png
+https://heyanon.xyz/magic/qa6lip,Apprentice_77.png
+https://heyanon.xyz/magic/z2jodd,Apprentice_78.png
+https://heyanon.xyz/magic/rwaxgb,Apprentice_79.png
+https://heyanon.xyz/magic/162jxi,Apprentice_80.png
+https://heyanon.xyz/magic/3e88rv,Apprentice_81.png
+https://heyanon.xyz/magic/iimwlh,Apprentice_82.png
+https://heyanon.xyz/magic/k5thpz,Apprentice_83.png
+https://heyanon.xyz/magic/6dgecv,Apprentice_84.png
+https://heyanon.xyz/magic/8j3fyl,Apprentice_85.png
+https://heyanon.xyz/magic/rywdky,Apprentice_86.png
+https://heyanon.xyz/magic/7z9hhr,Apprentice_87.png
+https://heyanon.xyz/magic/gv9wb4,Apprentice_88.png
+https://heyanon.xyz/magic/awk4qd,Apprentice_89.png
+https://heyanon.xyz/magic/31j89f,Apprentice_90.png
+https://heyanon.xyz/magic/hxq7pi,Apprentice_91.png
+https://heyanon.xyz/magic/uzu4v5,Apprentice_92.png
+https://heyanon.xyz/magic/2e4ekh,Apprentice_93.png
+https://heyanon.xyz/magic/cfzfpl,Apprentice_94.png
+https://heyanon.xyz/magic/e2dc91,Apprentice_95.png
+https://heyanon.xyz/magic/ah1i7q,Apprentice_96.png
+https://heyanon.xyz/magic/wp8f8x,Apprentice_97.png
+https://heyanon.xyz/magic/1gjchr,Apprentice_98.png
+https://heyanon.xyz/magic/yz80vg,Apprentice_99.png
