+++ data/sbc/sbcSecrets2.csv
@@ -1,0 +1,100 @@
+https://heyanon.xyz/magic/qe1d13,Apprentice_100.png
+https://heyanon.xyz/magic/x6jage,Apprentice_101.png
+https://heyanon.xyz/magic/h4pnkg,Apprentice_102.png
+https://heyanon.xyz/magic/32fqpg,Apprentice_103.png
+https://heyanon.xyz/magic/k3srgl,Apprentice_104.png
+https://heyanon.xyz/magic/7s6of4,Apprentice_105.png
+https://heyanon.xyz/magic/c8upt7,Apprentice_106.png
+https://heyanon.xyz/magic/olmrxp,Apprentice_107.png
+https://heyanon.xyz/magic/8o8zoa,Apprentice_108.png
+https://heyanon.xyz/magic/vorwx5,Apprentice_109.png
+https://heyanon.xyz/magic/bty4o9,Apprentice_110.png
+https://heyanon.xyz/magic/eh69rd,Apprentice_111.png
+https://heyanon.xyz/magic/nd8pz2,Apprentice_112.png
+https://heyanon.xyz/magic/wnzb5x,Apprentice_113.png
+https://heyanon.xyz/magic/jzz29o,Apprentice_114.png
+https://heyanon.xyz/magic/9bp5yr,Apprentice_115.png
+https://heyanon.xyz/magic/4nj2bd,Apprentice_116.png
+https://heyanon.xyz/magic/bg0sjx,Apprentice_117.png
+https://heyanon.xyz/magic/02168q,Apprentice_118.png
+https://heyanon.xyz/magic/dctlk4,Apprentice_119.png
+https://heyanon.xyz/magic/us9exh,Apprentice_120.png
+https://heyanon.xyz/magic/zomtsz,Apprentice_121.png
+https://heyanon.xyz/magic/g1ekg5,Apprentice_122.png
+https://heyanon.xyz/magic/8zz19z,Apprentice_123.png
+https://heyanon.xyz/magic/091u3j,Apprentice_124.png
+https://heyanon.xyz/magic/pezoxy,Apprentice_125.png
+https://heyanon.xyz/magic/huypja,Apprentice_126.png
+https://heyanon.xyz/magic/4bwalr,Apprentice_127.png
+https://heyanon.xyz/magic/7n2ls9,Apprentice_128.png
+https://heyanon.xyz/magic/fxhvif,Apprentice_129.png
+https://heyanon.xyz/magic/513ddn,Apprentice_130.png
+https://heyanon.xyz/magic/5j3eqy,Apprentice_131.png
+https://heyanon.xyz/magic/it5hin,Apprentice_132.png
+https://heyanon.xyz/magic/y4yxo2,Apprentice_133.png
+https://heyanon.xyz/magic/0by2u9,Apprentice_134.png
+https://heyanon.xyz/magic/0h0v6q,Apprentice_135.png
+https://heyanon.xyz/magic/d1k41j,Apprentice_136.png
+https://heyanon.xyz/magic/5kcmri,Apprentice_137.png
+https://heyanon.xyz/magic/48rhcm,Apprentice_138.png
+https://heyanon.xyz/magic/h05uxt,Apprentice_139.png
+https://heyanon.xyz/magic/p4opcf,Apprentice_140.png
+https://heyanon.xyz/magic/mhqlwk,Apprentice_141.png
+https://heyanon.xyz/magic/sdzuaw,Apprentice_142.png
+https://heyanon.xyz/magic/ktzhl5,Apprentice_143.png
+https://heyanon.xyz/magic/i32jlb,Apprentice_144.png
+https://heyanon.xyz/magic/of6fbw,Apprentice_145.png
+https://heyanon.xyz/magic/xmvkck,Apprentice_146.png
+https://heyanon.xyz/magic/i0erv0,Apprentice_147.png
+https://heyanon.xyz/magic/0h2077,Apprentice_148.png
+https://heyanon.xyz/magic/hs3owj,Apprentice_149.png
+https://heyanon.xyz/magic/rswloq,Wizard_150.png
+https://heyanon.xyz/magic/3fvfnj,Wizard_151.png
+https://heyanon.xyz/magic/h5a279,Wizard_152.png
+https://heyanon.xyz/magic/406jw7,Wizard_153.png
+https://heyanon.xyz/magic/3c1tv6,Wizard_154.png
+https://heyanon.xyz/magic/4s4hlm,Wizard_155.png
+https://heyanon.xyz/magic/nryqms,Wizard_156.png
+https://heyanon.xyz/magic/4eb1i5,Wizard_157.png
+https://heyanon.xyz/magic/7ss0yk,Wizard_158.png
+https://heyanon.xyz/magic/zs2tu7,Wizard_159.png
+https://heyanon.xyz/magic/qg2d88,Wizard_160.png
+https://heyanon.xyz/magic/xjlzf1,Wizard_161.png
+https://heyanon.xyz/magic/l87mi6,Wizard_162.png
+https://heyanon.xyz/magic/g7nllg,Wizard_163.png
+https://heyanon.xyz/magic/cg52bm,Wizard_164.png
+https://heyanon.xyz/magic/8api4v,Wizard_165.png
+https://heyanon.xyz/magic/fka5fd,Wizard_166.png
+https://heyanon.xyz/magic/8gdurv,Wizard_167.png
+https://heyanon.xyz/magic/pac9nk,Wizard_168.png
+https://heyanon.xyz/magic/wqscos,Wizard_169.png
+https://heyanon.xyz/magic/fjnx4t,Wizard_170.png
+https://heyanon.xyz/magic/rgu7nb,Wizard_171.png
+https://heyanon.xyz/magic/9g6k0h,Wizard_172.png
+https://heyanon.xyz/magic/ewcar2,Wizard_173.png
+https://heyanon.xyz/magic/8k8oan,Wizard_174.png
+https://heyanon.xyz/magic/2oxuh5,Grand wizard_175.png
+https://heyanon.xyz/magic/6mao73,Grand wizard_176.png
+https://heyanon.xyz/magic/l52idq,Grand wizard_177.png
+https://heyanon.xyz/magic/h3zmgi,Grand wizard_178.png
+https://heyanon.xyz/magic/hw4mni,Grand wizard_179.png
+https://heyanon.xyz/magic/vl26xu,Grand wizard_180.png
+https://heyanon.xyz/magic/kp84ce,Grand wizard_181.png
+https://heyanon.xyz/magic/jr35a2,Grand wizard_182.png
+https://heyanon.xyz/magic/zze5fm,Grand wizard_183.png
+https://heyanon.xyz/magic/65ils6,Grand wizard_184.png
+https://heyanon.xyz/magic/iskui4,Grand wizard_185.png
+https://heyanon.xyz/magic/0kabqr,Grand wizard_186.png
+https://heyanon.xyz/magic/8jqj59,Grand wizard_187.png
+https://heyanon.xyz/magic/1isbe4,Grand wizard_188.png
+https://heyanon.xyz/magic/fjpnc0,Grand wizard_189.png
+https://heyanon.xyz/magic/yb6pt0,Sorcerer_190.png
+https://heyanon.xyz/magic/rftuph,Sorcerer_191.png
+https://heyanon.xyz/magic/3yzn7s,Sorcerer_192.png
+https://heyanon.xyz/magic/o4i6en,Sorcerer_193.png
+https://heyanon.xyz/magic/kpbim6,Sorcerer_194.png
+https://heyanon.xyz/magic/zkrbcb,Sorcerer_195.png
+https://heyanon.xyz/magic/hwuje0,Sorcerer_196.png
+https://heyanon.xyz/magic/ox5rv7,Sorcerer_197.png
+https://heyanon.xyz/magic/ctbmac,Sorcerer_198.png
+https://heyanon.xyz/magic/7llts6,Sorcerer_199.png
